fix: start cleaning up `detect-known-trace.ts`
